+++ Sources/SwiftDisc/Models/Emoji.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 public struct Emoji: Codable, Hashable {
-    public let id: Snowflake?
+    public let id: EmojiID?
     public let name: String?
     public let roles: [RoleID]?
     public let user: User?