+++ Sources/SwiftDisc/Models/GuildWidgetSettings.swift
@@ -2,5 +2,5 @@ import Foundation
 
 public struct GuildWidgetSettings: Codable, Hashable {
     public let enabled: Bool
-    public let channel_id: Snowflake?
+    public let channel_id: ChannelID?
 }