+++ Sources/SwiftDisc/Models/PartialGuild.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public struct PartialGuild: Codable, Hashable {
+    public let id: GuildID
+    public let name: String
+}