+++ Sources/SwiftDisc/Models/ScheduledEventUser.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 public struct GuildScheduledEventUser: Codable, Hashable {
-    public let guild_scheduled_event_id: Snowflake
+    public let guild_scheduled_event_id: GuildScheduledEventID
     public let user: User
     public let member: GuildMember?
 }