+++ Sources/SwiftDisc/Models/Webhook.swift
@@ -1,10 +1,10 @@
 import Foundation
 
 public struct Webhook: Codable, Hashable {
-    public let id: Snowflake
+    public let id: WebhookID
     public let type: Int
-    public let channel_id: Snowflake?
-    public let guild_id: Snowflake?
+    public let channel_id: ChannelID?
+    public let guild_id: GuildID?
     public let name: String?
     public let token: String?
 }